// src/api/generated/models/ticketCommentCreateRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */

/**
 * Create one immutable ticket comment event.
 */
export interface TicketCommentCreateRequest {
  /**
   * @minLength 1
   * @maxLength 2000
   */
  text: string
  is_internal?: boolean
}